schedule float to int

+++ Assets/Scripts/Planning/SS_Schedule.cs
@@ -4,9 +4,9 @@ using System;
 [Serializable]
 public class SS_Schedule
 {
-    [SerializeField, Range(0, 24)] float startTime = 0;
-    [SerializeField, Range(0, 24)] float endTime = 0;
+    [SerializeField, Range(0, 24)] int startTime = 0;
+    [SerializeField, Range(0, 24)] int endTime = 0;
 
-    public float StartTime => startTime;
-    public float EndTime => endTime;
+    public int StartTime => startTime;
+    public int EndTime => endTime;
 }
